Add keyboard shortcut to fall back to usable config
The code for creating the shortcut is (almost) at the very top of
rc.lua, so if anything going wrong after that results in an
unusable desktop enviroment (i. e. one from which it is impossible to
restore a working rc.lua), the script to fall back to a usable config
can be launched with the keyboard combination `Super + Alt + p`.
Note that I haven't actually tested the scenario where this 'panic-mode'
is needed.